update fairness module
