+++ cider/fairness.py
@@ -1,91 +1,77 @@
-from datastore import DataStore, DataType
-from helpers.utils import make_dir
-from helpers.plot_utils import clean_plot
-from matplotlib.collections import PatchCollection  # type: ignore[import]
-import matplotlib.pyplot as plt  # type: ignore[import]
-import numpy as np
-import pandas as pd
-from pandas import DataFrame as PandasDataFrame
-from scipy.stats import f_oneway  # type: ignore[import]
-import seaborn as sns  # type: ignore[import]
-from typing import Dict, Mapping, List, Optional, Union
+"""
+Evaluates fairness of a machine learning module across a characteristic 
+(whether or not the machine learning module discriminates across different groups in that characteristic).
+"""
+from numpy import character
+from box import Box
+import yaml
+from helpers.utils import *
+from helpers.plot_utils import *
+from helpers.io_utils import *
+from helpers.ml_utils import *
+from scipy.stats import f_oneway, chi2_contingency
+from datastore import *
+from sklearn.metrics import recall_score, precision_score
 
 
 class Fairness:
 
-    def __init__(self, datastore: DataStore, clean_folders: bool = False) -> None:
+    def __init__(self, datastore: DataStore, dataframe=None, clean_folders=False):
         self.cfg = datastore.cfg
         self.ds = datastore
-        self.outputs = datastore.outputs + 'fairness/'
 
-        # Prepare working directories, set color palette
-        make_dir(self.outputs, clean_folders)
+        data_path = self.cfg.path.data + self.cfg.path.file_names.fairness
+        self.data = pd.read_csv(data_path)
+        self.data['random'] = np.random.rand(len(self.data))
+
+        outputs = self.cfg.path.outputs
+        self.outputs = outputs
         self.default_colors = sns.color_palette('Set2', 100)
 
-        # Load data into datastore
-        data_type_map = {DataType.FAIRNESS: None}
-        self.ds.load_data(data_type_map=data_type_map)
+        # Unweighted data
+        self.unweighted_data = self.data.copy()
+        self.unweighted_data['weight'] = 1
 
-    def rank_residual(self, var1: str, var2: str,
-                      characteristic: str, weighted: bool = False) -> Mapping[str, np.ndarray]:
-        """
-        Returns the distribution of normalized rank-residuals across subgroups and targeting methods.
-        The rank-residual is a measure of whether certain groups are consistently ranked higher or lower than they
-        "should" be, as defined by Aiken et al. (2021).
+        # Weighted data
+        self.weighted_data = self.data.copy()
+        if 'weight' not in self.weighted_data.columns:
+            self.data['weight'] = 1
+        else:
+            self.weighted_data['weight'] = self.weighted_data['weight']/self.weighted_data['weight'].min()
+        self.weighted_data = pd.DataFrame(np.repeat(self.weighted_data.values, self.weighted_data['weight'], axis=0), columns=self.weighted_data.columns)\
+            .astype(self.unweighted_data.dtypes)
+    # --------------------------------
+    # SINGLULAR FUNCTIONS
+    # --------------------------------
 
-        Args:
-            var1: The name of the column containing data coming from the first targeting method; can also be the ground-
-            truth column.
-            var2: The name of the column containing data coming from the second targeting method.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-            of the corresponding column in the targeting dataset.
-            weighted: If True the weighted version of the dataset will be used.
-
-        Returns:
-            A dict mapping the characteristic's subgroups to their distributions of normalized rank-residuals.
-        """
-
-        data = self.ds.weighted_fairness if weighted else self.ds.unweighted_fairness
-
+    def rank_residual(self, var1, var2, characteristic, weighted=False):
+        
+        data = self.weighted_data if weighted else self.unweighted_data
+        
         # Get rankings according to var1 and var2
         data = data.sort_values([var1, 'random'], ascending=True)
         data['rank1'] = range(len(data))
         data = data.sort_values([var2, 'random'], ascending=True)
         data['rank2'] = range(len(data))
 
-        # Calculate normalized rank residual
-        data['rank_residual'] = (data['rank2'] - data['rank1']) / data['rank2'].max()
+        # Calcualte normalized rank residual
+        data['rank_residual'] = (data['rank2'] - data['rank1'])/data['rank2'].max()
 
         # Return rank residual distributions for each group
-        return {group: data[data[characteristic] == group]['rank_residual'].values.flatten()
-                for group in data[characteristic].unique()}
+        return {group: data[data[characteristic] == group]['rank_residual'].values.flatten() for group in data[characteristic].unique()}
 
-    def demographic_parity(self, var1: str, var2: str, characteristic: str,
-                           p: Union[float, int], weighted: bool = False
-                           ) -> Dict[str, Dict[str, float]]:
+    def demographic_parity(self, var1, var2, characteristic, p, weighted=False):
         """
-        Compares the proportion of each subgroup living in poverty (below the pth percentile in terms of consumption)
-        to the proportion of each subgroup that is targeted (below the pth percentile in terms of the proxy poverty
-        measure used for targeting) to produce a measure of demographic parity. This is defined as:
-        DP = (TP + FP)/N - (TP + FN)/N
-        where TP stands for true positives, and so forth.
-
-        Args:
-            var1: The name of the column containing data coming from the first targeting method; can also be the ground-
-            truth column.
-            var2: The name of the column containing data coming from the second targeting method.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-                of the corresponding column in the targeting dataset.
-            p: The percentile below which users will be considered as poor.
-            weighted: If True the weighted version of the dataset will be used.
-
-        Returns: A dict mapping the characteristic's subgroups to their poverty shares and demographic parities.
+        var1 = groundtruth
+        var2 = proxies. The thing that you're evaluating the fairness of.
+        characteristic = how the group is divided
+        p = target the top p percent
         """
 
-        data = self.ds.weighted_fairness if weighted else self.ds.unweighted_fairness
+        data = self.weighted_data if weighted else self.unweighted_data
 
         # Simulate targeting by var1 and var2
-        num_ones = int((p / 100) * len(data))
+        num_ones = int((p/100)*len(data))
         num_zeros = len(data) - num_ones
         targeting_vector = np.concatenate([np.ones(num_ones), np.zeros(num_zeros)])
         data = data.sort_values([var1, 'random'], ascending=True)
@@ -94,103 +80,118 @@ class Fairness:
         data['targeted_var2'] = targeting_vector
 
         # Get demographic parity and poverty share for each group
-        results: Dict[str, Dict[str, float]] = {}
+        results = {}
         for group in data[characteristic].unique():
             subset = data[data[characteristic] == group]
             results[group] = {}
-            results[group]['poverty_share'] = 100 * subset['targeted_var1'].mean()
-            results[group]['demographic_parity'] = 100 * subset['targeted_var2'].mean() - \
-                                                   100 * subset['targeted_var1'].mean()
+            results[group]['poverty_share'] = 100*subset['targeted_var1'].mean()
+            results[group]['demographic_parity'] = 100*subset['targeted_var2'].mean() - 100*subset['targeted_var1'].mean()
         return results
-
-    def rank_residuals_plot(self, groundtruth: str, proxies: List[str], characteristic: str,
-                            weighted: bool = False, colors: Optional[sns.color_palette] = None) -> None:
+    
+    def independence(self, var1, var2, characteristic, p, weighted=False):
         """
-        Plots the rank-residuals, as box plots, between a list of proxies and the ground-truth data, disaggregated by
-        a (demographic) characteristic.
+        Performs the chi-squared independence test between the proxy variable and the sensitive characterisrtic
 
-        Args:
-            groundtruth: The name of the groundtruth column in the fairness dataset.
-            proxies: The list of targeting methods to be compared against each other; each name should have a
-                corresponding column in the fairness dataset.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-                of the corresponding column in the targeting dataset.
-            weighted: If True the weighted version of the dataset will be used.
-            colors: The color palette to use when plotting.
+        var2 -- the proxy variable
+        characteristic -- sensitive characteristic
+
         """
+        data = self.weighted_data if weighted else self.unweighted_data
 
-        data = self.ds.weighted_fairness if weighted else self.ds.unweighted_fairness
+        # Simulate targeting by var1 and var2
+        num_ones = int((p/100)*len(data))
+        num_zeros = len(data) - num_ones
+        targeting_vector = np.concatenate([np.ones(num_ones), np.zeros(num_zeros)])
+        data = data.sort_values([var2, 'random'], ascending=True)
+        data['targeted_var2'] = targeting_vector
 
-        # Set up figure
-        fig, ax = plt.subplots(1, len(proxies), figsize=(20, 7), sharey=True)
-        max_resid = 0
+        # Get targeted percent for each group
+        results = {}
+        for group in data[characteristic].unique():
+            subset = data[data[characteristic] == group]
+            results[group] = subset['targeted_var2'].mean()
 
-        # Set up color palette
-        if colors is None:
-            colors = self.default_colors
+        # Create a pivot table between the characteristic and the proxy
+        pivot = data.pivot_table(index=characteristic, columns='targeted_var2', aggfunc='count', fill_value=0).iloc[:, 0:2]
 
-        for i, proxy in enumerate(proxies):
-
-            # Obtain rank residual distribution for each group
-            distributions: List[Mapping[str, np.ndarray]] = []
-            groups: List[int] = []
-            results = self.rank_residual(groundtruth, proxy, characteristic, weighted=weighted)
-            for g, group in enumerate(sorted(data[characteristic].unique())):
-                distribution = list(results[group])
-                distributions = distributions + distribution
-                groups = groups + [group for _ in range(len(distribution))]
-
-            # Combine rank residual distributions for all groups into dataframe
-            results_df = pd.DataFrame([distributions, groups]).T
-            results_df.columns = ['Normalized Rank Residual', characteristic]
-            results_df = results_df.sort_values(characteristic, ascending=True)
-
-            # Update range for plotting
-            if np.abs(results_df['Normalized Rank Residual'].max()) > max_resid:
-                max_resid = np.abs(results_df['Normalized Rank Residual'].max())
-            if np.abs(results_df['Normalized Rank Residual'].min()) > max_resid:
-                max_resid = np.abs(results_df['Normalized Rank Residual'].min())
-
-            # Create boxplot
-            sns.boxplot(data=results_df, x='Normalized Rank Residual', y=characteristic, orient='h', ax=ax[i],
-                        showfliers=False,
-                        palette={group: colors[g] for g, group in enumerate(sorted(data[characteristic].unique()))})
-            ax[i].axvline(0, color='grey', dashes=[1, 1])
-
-            # Clean up subplot
-            ax[i].set_ylabel('')
-            ax[i].set_xlabel('Error')
-            ax[i].set_title(proxy, fontsize='large')
-
-        # Clean up overall plot
-        for i in range(len(ax)):
-            ax[i].set_xlim(-max_resid, max_resid)
-            clean_plot(ax[i])
-        plt.tight_layout(rect=[0, 0, 1, .92])
-        plt.suptitle('Rank Fairness: ' + characteristic, fontsize='x-large')
-
-        # Save and show
-        plt.savefig(self.outputs + '/rank_residuals_plot_' + characteristic + '.png', dpi=400)
-        plt.show()
-
-    def rank_residuals_table(self, groundtruth: str, proxies: List[str], characteristic: str,
-                             weighted: bool = False) -> PandasDataFrame:
+        # Run independence test
+        chi2, p, dof, ex = chi2_contingency(pivot)
+        results['ind_pval'] = p
+        return results
+    
+    def recall_per_group(self, var1, var2, characteristic, p, weighted=False):
         """
-        Returns a dataframe with the rank-residuals of all sub-groups part of 'characteristic', and for all proxies
-        specified by 'proxies'. It also performs an ANOVA test.
-
-        Args:
-            groundtruth: The name of the groundtruth column in the fairness dataset.
-            proxies: The list of targeting methods to be compared against each other; each name should have a
-                corresponding column in the fairness dataset.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-                of the corresponding column in the targeting dataset.
-            weighted: If True the weighted version of the dataset will be used.
-
-        Returns: The pandas dataframe of rank-residuals and ANOVA F-stat and p-value.
+        var1 = groundtruth
+        var2 = proxies. The thing that you're evaluating the fairness of.
+        characteristic = how the group is divided
+        p = target the top p percent
         """
 
-        data = self.ds.weighted_fairness if weighted else self.ds.unweighted_fairness
+        data = self.weighted_data if weighted else self.unweighted_data
+
+        # Simulate targeting by var1 and var2
+        num_ones = int((p/100)*len(data))
+        num_zeros = len(data) - num_ones
+        targeting_vector = np.concatenate([np.ones(num_ones), np.zeros(num_zeros)])
+        data = data.sort_values([var1, 'random'], ascending=True)
+        data['targeted_var1'] = targeting_vector
+        data = data.sort_values([var2, 'random'], ascending=True)
+        data['targeted_var2'] = targeting_vector
+
+        # Get recall for each group
+        results = {}
+        for group in data[characteristic].unique():
+            subset = data[data[characteristic] == group]
+            results[group] = recall_score(subset['targeted_var1'], subset['targeted_var2'], zero_division=0)
+        
+        # Create a pivot table between true positives and false negatives
+        pivot = data[data['targeted_var1']==1].pivot_table(index=characteristic, columns='targeted_var2', aggfunc='count', fill_value=0).iloc[:, 0:2]
+
+        # Run independence test
+        chi2, p, dof, ex = chi2_contingency(pivot)
+        results['ind_pval'] = p
+        return results
+    
+    def precision_per_group(self, var1, var2, characteristic, p, weighted=False):
+        """
+        var1 = groundtruth
+        var2 = proxies. The thing that you're evaluating the fairness of.
+        characteristic = how the group is divided
+        p = target the top p percent
+        """
+
+        data = self.weighted_data if weighted else self.unweighted_data
+
+        # Simulate targeting by var1 and var2
+        num_ones = int((p/100)*len(data))
+        num_zeros = len(data) - num_ones
+        targeting_vector = np.concatenate([np.ones(num_ones), np.zeros(num_zeros)])
+        data = data.sort_values([var1, 'random'], ascending=True)
+        data['targeted_var1'] = targeting_vector
+        data = data.sort_values([var2, 'random'], ascending=True)
+        data['targeted_var2'] = targeting_vector
+
+        # Get demographic parity and poverty share for each group
+        results = {}
+        for group in data[characteristic].unique():
+            subset = data[data[characteristic] == group]
+            results[group] = precision_score(subset['targeted_var1'], subset['targeted_var2'], zero_division=0)
+
+        # Create a pivot table between true and false positives
+        pivot = data[data['targeted_var2']==1].pivot_table(index=characteristic, columns='targeted_var1', aggfunc='count', fill_value=0).iloc[:, 0:2]
+
+        # Run independence test
+        chi2, p, dof, ex = chi2_contingency(pivot)
+        results['ind_pval'] = p
+        return results
+    
+    # --------------------------------
+    # TABLE GENERATION FUNCTIONS
+    # --------------------------------
+
+    def rank_residuals_table(self, groundtruth, proxies, characteristic, weighted=False):
+
+        data = self.weighted_data if weighted else self.unweighted_data
 
         # Set up table
         table = pd.DataFrame()
@@ -212,83 +213,149 @@ class Fairness:
 
             # Add column to table
             table[proxy] = column
-
+        
+        table.set_index(characteristic, inplace=True)
+        
         # Save and return
         table.to_csv(self.outputs + '/rank_residuals_table_' + characteristic + '.png', index=False)
         return table
 
-    def demographic_parity_table(self, groundtruth: str, proxies: List[str], characteristic: str,
-                                 p: Union[float, int], weighted: bool = False, format_table: bool = True
-                                 ) -> PandasDataFrame:
-        """
-        Returns a dataframe with the demographic parity vales of all sub-groups part of 'characteristic', and for all
-        proxies specified by 'proxies'.
+    def demographic_parity_table(self, groundtruth, proxies, characteristic, p, weighted=False, format_table=True):
 
-        Args:
-            groundtruth: The name of the groundtruth column in the fairness dataset.
-            proxies: The list of targeting methods to be compared against each other; each name should have a
-                corresponding column in the fairness dataset.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-                of the corresponding column in the targeting dataset.
-            p: The percentile below which users will be considered as poor.
-            weighted: If True the weighted version of the dataset will be used.
-            format_table: If true, provide only two significant decimal digits.
-
-        Returns: The pandas dataframe with DP values, each group's share of the population, and its share of the target
-        population.
-        """
-
-        data = self.ds.weighted_fairness if weighted else self.ds.unweighted_fairness
+        data = self.weighted_data if weighted else self.unweighted_data
         data['count'] = 1
 
         # Set up table
-        table = pd.DataFrame()
+        table=pd.DataFrame()
         groups = sorted(data[characteristic].unique())
         table[characteristic] = groups
 
         # Get share of population for each group
         population_shares = data.groupby(characteristic).agg('count')
-        table["Group's share of population"] = 100 * (population_shares['count'].values.flatten() / len(data))
+        table["Group's share of population"] = 100*(population_shares['count'].values.flatten()/len(data))
 
         # Get demographic parity and poverty share for each group
         for proxy in proxies:
             groups = sorted(data[characteristic].unique())
             results = self.demographic_parity(groundtruth, proxy, characteristic, p, weighted=weighted)
             if proxy == proxies[0]:
-                table["Share of Group in Target Population"] = [results[group]['poverty_share'] for group in groups]
+                table["Share of Group in Target Population"] = [results[group]['poverty_share'] for group in groups] 
             table[proxy] = [results[group]['demographic_parity'] for group in groups]
-
+        
         # Clean up and return table
         if format_table:
-            table["Group's share of population"] = table["Group's share of population"].apply(
-                lambda x: ('%.2f' % x) + '%')
-            table["Share of Group in Target Population"] = table["Share of Group in Target Population"].apply(
-                lambda x: ('%.2f' % x) + '%')
+            table["Group's share of population"] = table["Group's share of population"].apply(lambda x: ('%.2f' % x) + '%')
+            table["Share of Group in Target Population"] = table["Share of Group in Target Population"].apply(lambda x: ('%.2f' % x) + '%')
+            table.set_index(characteristic, inplace=True)
 
         # Save and return
         table.to_csv(self.outputs + '/demographic_parity_table_' + characteristic + '_' + str(p) + '%.png', index=False)
         return table
 
-    def demographic_parity_plot(self, groundtruth: str, proxies: List[str], characteristic: str,
-                                p: Union[float, int], weighted: bool = False) -> None:
-        """
-        Plots the demographic parity values, with proxies on the x-axis and subgroups on the y-axis. The magnitude of
-        the value determines the corresponding circle's radius, while its sign the circle's color (red for positive,
-        blue for negative).
+    def create_table(self, function, name, groundtruth, proxies, characteristic, p, weighted=False, format_table=True):
+        data = self.weighted_data if weighted else self.unweighted_data
+        data['count'] = 1
 
-        Args:
-            groundtruth: The name of the groundtruth column in the fairness dataset.
-            proxies: The list of targeting methods to be compared against each other; each name should have a
-                corresponding column in the fairness dataset.
-            characteristic: The (demographic) characteristic by which to disaggregate results - this should be the name
-                of the corresponding column in the targeting dataset.
-            p: The percentile below which users will be considered as poor.
-            weighted: If True the weighted version of the dataset will be used.
-        """
+        # Set up table
+        table=pd.DataFrame()
+        groups = sorted(data[characteristic].unique())
+        groups.append('ind_pval')
+        table[characteristic] = groups
 
-        # Get demographic parity table, set up parameters for grid 
-        table = self.demographic_parity_table(groundtruth, proxies, characteristic, p, weighted=weighted,
-                                              format_table=False)
+        # Get share of population for each group
+        population_shares = data.groupby(characteristic).agg('count')
+        table["Group's share of population"] = list(100*(population_shares['count'].values.flatten()/len(data))) + [0]
+
+        # Get demographic parity and poverty share for each group
+        for proxy in proxies:
+            groups = sorted(data[characteristic].unique())
+            results = function(groundtruth, proxy, characteristic, p, weighted=weighted)
+            table[proxy] = [results[group] for group in groups] + [results['ind_pval']]
+        
+        # Clean up and return table
+        if format_table:
+            table["Group's share of population"] = table["Group's share of population"].apply(lambda x: ('%.2f' % x) + '%')
+            table.iloc[-1, 1] = ''
+            table.iloc[:, 2:] = table.iloc[:, 2:].applymap('{:.4f}'.format)
+            table.set_index(characteristic, inplace=True)
+
+        # Save and return
+        table.to_csv(self.outputs + f'/{name}_table_' + characteristic + '_' + str(p) + '%.png', index=False)
+        return table
+    
+    def independence_table(self, *args, **kwargs):
+        return self.create_table(self.independence, 'independence', *args, **kwargs)
+    
+    def recall_table(self, *args, **kwargs):
+        return self.create_table(self.recall_per_group, 'recall', *args, **kwargs)
+    
+    def precision_table(self, *args, **kwargs):
+        return self.create_table(self.precision_per_group, 'precision', *args, **kwargs)
+    
+    # --------------------------------
+    # PLOTTING FUNCTIONS
+    # --------------------------------
+
+    def rank_residuals_plot(self, groundtruth, proxies, characteristic, weighted=False, colors=None):
+
+        data = self.weighted_data if weighted else self.unweighted_data
+
+        # Set up figure
+        fig, ax = plt.subplots(1, len(proxies), figsize=(20, 7), sharey=True)
+        max_resid = 0
+
+        # Set up color palette
+        if colors is None:
+            colors = self.default_colors
+
+        for i, proxy in enumerate(proxies):
+
+            # Obtain rank residual distribution for each group
+            distributions, groups = [], []
+            results = self.rank_residual(groundtruth, proxy, characteristic, weighted=weighted)
+            for g, group in enumerate(sorted(data[characteristic].unique())):
+                distribution = list(results[group])
+                distributions = distributions + distribution
+                groups = groups + [group for _ in range(len(distribution))]
+
+            # Combine rank residual distributions for all groups into dataframe
+            results = pd.DataFrame([distributions, groups]).T
+            results.columns = ['Normalized Rank Residual', characteristic]
+            results = results.sort_values(characteristic, ascending=True)
+
+            # Update range for plotting
+            if np.abs(results['Normalized Rank Residual'].max()) > max_resid:
+                max_resid = np.abs(results['Normalized Rank Residual'].max())
+            if np.abs(results['Normalized Rank Residual'].min()) > max_resid:
+                max_resid = np.abs(results['Normalized Rank Residual'].min())
+
+            # Create boxplot
+            sns.boxplot(data=results, x='Normalized Rank Residual', y=characteristic, orient='h', ax=ax[i], showfliers=False, 
+                palette={group: colors[g] for g, group in enumerate(sorted(data[characteristic].unique()))})
+            ax[i].axvline(0, color='grey', dashes=[1, 1])
+
+            # Clean up subplot
+            ax[i].set_ylabel('')
+            ax[i].set_xlabel('Error')
+            ax[i].set_title(proxy, fontsize='large')
+        
+        # Clean up overall plot
+        for i in range(len(ax)):
+            ax[i].set_xlim(-max_resid, max_resid)
+            clean_plot(ax[i])
+        plt.tight_layout(rect=[0, 0, 1, .92])
+        plt.suptitle('Rank Fairness: ' + characteristic, fontsize='x-large')
+
+        # Save and show
+        plt.savefig(self.outputs + '/rank_residuals_plot_' + characteristic + '.png', dpi=400)
+        plt.show()
+
+    def create_plot(self, table, name, groundtruth, proxies, characteristic, p, weighted=False):
+        
+        # Remove p-values and set up parameters for grid 
+        dp = name == 'demographic_parity'
+        title = name.replace('_', ' ').title()
+        table = table if dp else table.iloc[:-1, :]
         data = table[proxies]
         keys = list(data.keys())
         N = len(table)
@@ -297,51 +364,59 @@ class Fairness:
         # Format labels for y axis
         ylabels = []
         for i in range(len(table)):
-            ylabels.append('{}\n{}% of Population\n{}% in Target'
-                           .format(table.iloc[i][characteristic], int(table.iloc[i]["Group's share of population"]),
-                                   int(table.iloc[i]['Share of Group in Target Population'])))
+            if dp:
+                ylabels.append('{}\n{}% of Population\n{}% in Target'\
+                        .format(table.iloc[i][characteristic], int(table.iloc[i]["Group's share of population"]), 
+                        int(table.iloc[i]['Share of Group in Target Population'])))
+            else:
+                ylabels.append('{}\n{}% of Population'\
+                            .format(table.iloc[i][characteristic], int(table.iloc[i]["Group's share of population"])))
         ylabels = ylabels[::-1]
         xlabels = keys
 
         # Circles
         x, y = np.meshgrid(np.arange(M), np.arange(N))
-        radius = 15
-        s: List[List[float]] = [[] for _ in range(len(keys))]
+        s = [[] for i in range(len(keys))]
         for i in range(len(keys)):
             for j in range(len(data[keys[i]])):
                 s[i].append(data[keys[i]][j])
         arr = np.array(s).transpose()
         new_list = []
-        for i in range(arr.shape[0] - 1, -1, -1):
+        for i in range(arr.shape[0]-1,-1,-1):
             new_list.append(list(arr[i]))
         s = new_list
+        # S contains the recall values, from bottom to top, from left to right
 
         # set up figure
-        fig = plt.figure(figsize=(2.7 * len(proxies), 2.5 * len(table)))
+        fig = plt.figure(figsize=(2.7*len(proxies), 2.5*len(table))) 
         ax = fig.add_axes([0, 0, 1, 1])
-        ax.set_title('Demographic Parity: ' + characteristic, pad=85, fontsize='large')
+        plot_name = f'{title} per Group: ' if not dp else f'{title}: '
+        ax.set_title(plot_name + characteristic, pad=85, fontsize='large')
 
         # More circles
-        R = np.array([np.array(row) for row in s])
+        s=np.array([np.array(row) for row in s])
+        R = s
         c = R
-        R = np.log(np.abs(R), where=np.abs(R) > 0) / 10
-        circles = [plt.Circle((j, i), radius=r) for r, j, i in zip(R.flatten(), x.flatten(), y.flatten())]
+        R = np.log(np.abs(R) + 1)/10 if dp else R**.5 / 3
+        circles = [plt.Circle((j,i), radius=r) for r, j, i in zip(R.flatten(), x.flatten(), y.flatten())]
         col = PatchCollection(circles, array=c.flatten(), cmap="RdBu_r", edgecolor='grey', linewidth=2)
-        col.set_clim(vmin=-20, vmax=20)
-        # math.log(abs(r)) / 10
+        if dp:
+            col.set_clim(vmin=-20, vmax=20)
+        else:
+            col.set_clim(vmin=0, vmax=1)
 
         # Set up ticks and labels
         ax.add_collection(col)
         ax.set(xticks=np.arange(M), yticks=np.arange(N),
-               xticklabels=xlabels, yticklabels=ylabels)
-        ax.set_xticks(np.arange(M + 1) - 0.5, minor=True)
-        ax.set_yticks(np.arange(N + 1) - 0.5, minor=True)
+            xticklabels=xlabels, yticklabels=ylabels)
+        ax.set_xticks(np.arange(M+1)-0.5, minor=True)
+        ax.set_yticks(np.arange(N+1)-0.5, minor=True)
         ax.xaxis.tick_top()
 
         # Colorbar
-        cbar = fig.colorbar(col, fraction=0.03, pad=0.05, )
+        cbar = fig.colorbar(col, fraction=0.03, pad=0.05,)
         cbar.outline.set_edgecolor('white')
-        cbar.ax.set_ylabel('Percentage Point Difference', labelpad=20)
+        cbar.ax.set_ylabel(f'{title}', labelpad=20)
 
         # Final touches on figure
         ax.spines['top'].set_visible(False)
@@ -349,7 +424,20 @@ class Fairness:
         ax.spines['bottom'].set_visible(False)
         ax.spines['left'].set_visible(False)
         ax.tick_params(axis='both', which='both', length=0)
+        plt.tight_layout()
 
         # Save and show
-        plt.savefig(self.outputs + '/demographic_parity_plot_' + characteristic + '_' + str(p) + '%.png')
+        plt.savefig(self.outputs + f'/{name}_plot_' + characteristic + '_' + str(p) + '%.png')
         plt.show()
+    
+    def demographic_parity_plot(self, *args, **kwargs):
+        self.create_plot(self.demographic_parity_table(*args, **kwargs, format_table=False), 'demographic_parity', *args, **kwargs)
+
+    def independence_plot(self, *args, **kwargs):
+        self.create_plot(self.independence_table(*args, **kwargs, format_table=False), 'independence', *args, **kwargs)
+    
+    def recall_plot(self, *args, **kwargs):
+        self.create_plot(self.recall_table(*args, **kwargs, format_table=False), 'recall', *args, **kwargs)
+    
+    def precision_plot(self, *args, **kwargs):
+        self.create_plot(self.precision_table(*args, **kwargs, format_table=False), 'precision', *args, **kwargs)
